[Refactor] refactor code into modules
